# server/internal/app/testdata/probe/openapi.yaml
openapi: 3.1.0
info: { title: The probe module's contract, version: '1' }
paths:
...
  /households:
    get: { operationId: getHouseholds, responses: { '200': { description: Listed }, default: { $ref: '#/components/responses/Problem' } } }
    post: { operationId: postHouseholds, requestBody: { $ref: '#/components/requestBodies/Any' }, responses: { '201': { description: Created }, default: { $ref: '#/components/responses/Problem' } } }
...
  /households/{household_id}/join-code:
    parameters: [{ $ref: '#/components/parameters/HouseholdId' }]
    post: { operationId: postHouseholdsByHouseholdIdJoinCode, responses: { '200': { description: Made }, default: { $ref: '#/components/responses/Problem' } } }
...
  /households/{household_id}/restriction:
    parameters: [{ $ref: '#/components/parameters/HouseholdId' }]
    post: { operationId: postHouseholdRestriction, requestBody: { required: false, content: { application/json: { schema: { type: object } } } }, responses: { '200': { description: Restricted }, default: { $ref: '#/components/responses/Problem' } } }
    delete: { operationId: deleteHouseholdRestriction, responses: { '200': { description: Lifted }, default: { $ref: '#/components/responses/Problem' } } }
  /households/{household_id}/members:
    parameters: [{ $ref: '#/components/parameters/HouseholdId' }]
    get: { operationId: getMembers, responses: { '200': { description: Listed }, default: { $ref: '#/components/responses/Problem' } } }
  /households/{household_id}/members/{user_id}:
    parameters: [{ $ref: '#/components/parameters/HouseholdId' }, { name: user_id, in: path, required: true, schema: { $ref: '#/components/schemas/Uuid' } }]
    get: { operationId: getMembersByUserId, responses: { '200': { description: Found }, default: { $ref: '#/components/responses/Problem' } } }
    patch: { operationId: patchMembersByUserId, requestBody: { $ref: '#/components/requestBodies/Any' }, responses: { '200': { description: Updated }, default: { $ref: '#/components/responses/Problem' } } }
    delete: { operationId: deleteMembersByUserId, responses: { '204': { description: Removed }, default: { $ref: '#/components/responses/Problem' } } }
  /households/{household_id}/ownership/transfer:
    parameters: [{ $ref: '#/components/parameters/HouseholdId' }]
    post: { operationId: postOwnershipTransfer, requestBody: { $ref: '#/components/requestBodies/Any' }, responses: { '200': { description: Promoted }, default: { $ref: '#/components/responses/Problem' } } }
  /households/{household_id}/invitations:
    parameters: [{ $ref: '#/components/parameters/HouseholdId' }]
    get: { operationId: getInvitations, responses: { '200': { description: Listed }, default: { $ref: '#/components/responses/Problem' } } }
    post: { operationId: postInvitations, requestBody: { $ref: '#/components/requestBodies/Any' }, responses: { '201': { description: Created }, default: { $ref: '#/components/responses/Problem' } } }
  /households/{household_id}/invitations/{invitation_id}:
    parameters: [{ $ref: '#/components/parameters/HouseholdId' }, { name: invitation_id, in: path, required: true, schema: { $ref: '#/components/schemas/Uuid' } }]
    delete: { operationId: deleteInvitationsByInvitationId, responses: { '204': { description: Revoked }, default: { $ref: '#/components/responses/Problem' } } }
  /households/{household_id}/invitations/{invitation_id}/resend:
    parameters: [{ $ref: '#/components/parameters/HouseholdId' }, { name: invitation_id, in: path, required: true, schema: { $ref: '#/components/schemas/Uuid' } }]
    post: { operationId: postInvitationsByInvitationIdResend, responses: { '202': { description: Sent }, default: { $ref: '#/components/responses/Problem' } } }
  /households/{household_id}/modules:
    parameters: [{ $ref: '#/components/parameters/HouseholdId' }]
    get: { operationId: getModules, responses: { '200': { description: Listed }, default: { $ref: '#/components/responses/Problem' } } }
  /households/{household_id}/modules/{module}:
    parameters: [{ $ref: '#/components/parameters/HouseholdId' }, { name: module, in: path, required: true, schema: { type: string } }]
    patch: { operationId: patchModulesByModule, requestBody: { $ref: '#/components/requestBodies/Any' }, responses: { '200': { description: Updated }, default: { $ref: '#/components/responses/Problem' } } }
  # The child profiles' routes the router always serves (item 11), declared as loosely as the tests need.
  /auth/child/profiles:
    post: { operationId: postAuthChildProfiles, security: [], requestBody: { $ref: '#/components/requestBodies/Any' }, responses: { '200': { description: Listed }, default: { $ref: '#/components/responses/Problem' } } }
  /auth/child/login:
    post: { operationId: postAuthChildLogin, security: [], requestBody: { $ref: '#/components/requestBodies/Any' }, responses: { '200': { description: Signed in }, default: { $ref: '#/components/responses/Problem' } } }
  /auth/graduation/confirm:
    post: { operationId: postAuthGraduationConfirm, security: [], requestBody: { $ref: '#/components/requestBodies/Any' }, responses: { '204': { description: Graduated }, default: { $ref: '#/components/responses/Problem' } } }
  /households/{household_id}/children:
    parameters: [{ $ref: '#/components/parameters/HouseholdId' }]
    post: { operationId: postChildren, requestBody: { $ref: '#/components/requestBodies/Any' }, responses: { '201': { description: Created }, default: { $ref: '#/components/responses/Problem' } } }
  /households/{household_id}/children/{user_id}/pin:
    parameters: [{ $ref: '#/components/parameters/HouseholdId' }, { name: user_id, in: path, required: true, schema: { $ref: '#/components/schemas/Uuid' } }]
    put: { operationId: putChildrenByUserIdPin, requestBody: { $ref: '#/components/requestBodies/Any' }, responses: { '204': { description: Set }, default: { $ref: '#/components/responses/Problem' } } }
  /households/{household_id}/children/{user_id}/unlock:
    parameters: [{ $ref: '#/components/parameters/HouseholdId' }, { name: user_id, in: path, required: true, schema: { $ref: '#/components/schemas/Uuid' } }]
    post: { operationId: postChildrenByUserIdUnlock, responses: { '204': { description: Unlocked }, default: { $ref: '#/components/responses/Problem' } } }
  /households/{household_id}/children/{user_id}/graduate:
    parameters: [{ $ref: '#/components/parameters/HouseholdId' }, { name: user_id, in: path, required: true, schema: { $ref: '#/components/schemas/Uuid' } }]
    post: { operationId: postChildrenByUserIdGraduate, requestBody: { $ref: '#/components/requestBodies/Any' }, responses: { '202': { description: Sent }, default: { $ref: '#/components/responses/Problem' } } }
  /households/{household_id}/children/{user_id}/avatar:
    parameters: [{ $ref: '#/components/parameters/HouseholdId' }, { name: user_id, in: path, required: true, schema: { $ref: '#/components/schemas/Uuid' } }]
    put: { operationId: putChildrenByUserIdAvatar, requestBody: { $ref: '#/components/requestBodies/Upload' }, responses: { '200': { description: Set }, default: { $ref: '#/components/responses/Problem' } } }
    delete: { operationId: deleteChildrenByUserIdAvatar, responses: { '200': { description: Removed }, default: { $ref: '#/components/responses/Problem' } } }
...
  /me/invitations:
    get: { operationId: getMeInvitations, responses: { '200': { description: Listed }, default: { $ref: '#/components/responses/Problem' } } }
  /me/invitations/{token}:
    parameters: [{ name: token, in: path, required: true, schema: { type: string } }]
    get: { operationId: getMeInvitationsByToken, security: [], responses: { '200': { description: Found }, default: { $ref: '#/components/responses/Problem' } } }
  /me/invitations/{token}/accept:
    parameters: [{ name: token, in: path, required: true, schema: { type: string } }]
    post: { operationId: postMeInvitationsByTokenAccept, responses: { '200': { description: Joined }, default: { $ref: '#/components/responses/Problem' } } }
  /me/invitations/{token}/decline:
    parameters: [{ name: token, in: path, required: true, schema: { type: string } }]
    post: { operationId: postMeInvitationsByTokenDecline, responses: { '204': { description: Declined }, default: { $ref: '#/components/responses/Problem' } } }
...
  /households/{household_id}/deletion:
    parameters: [{ $ref: '#/components/parameters/HouseholdId' }]
    post: { operationId: postDeletion, requestBody: { $ref: '#/components/requestBodies/Any' }, responses: { '202': { description: Scheduled }, default: { $ref: '#/components/responses/Problem' } } }
    delete: { operationId: deleteDeletion, responses: { '204': { description: Cancelled }, default: { $ref: '#/components/responses/Problem' } } }
...
components:
  parameters:
    HouseholdId: { name: household_id, in: path, required: true, schema: { $ref: '#/components/schemas/Uuid' } }
...
  requestBodies:
    Any: { required: true, content: { application/json: { schema: { type: object } } } }
    Upload: { required: true, content: { multipart/form-data: { schema: { type: object } } } }
  responses:
    Problem:
      description: A problem.
      content: { application/problem+json: { schema: { $ref: '#/components/schemas/Problem' } } }
  schemas:
    Uuid:
      type: string
      pattern: '^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{12}$'
...
    ProblemCode: { type: string }
    Problem:
      type: object
      required: [type, title, status, code]
      properties:
        type: { type: string }
        title: { type: string }
        status: { type: integer }
        code: { $ref: '#/components/schemas/ProblemCode' }
        request_id: { type: string }
...